zit: Don't use `git pull --rebase`
Tests one more time saving the day: adding a test for zit-il with branch
call showed that `git pull rebase` broke zit-up in the non-master branch
case.

So it was a bad idea, remove it. User is not supposed to change plugin
files directly anyway, and if he changed it he will be on its own.