zit: Require dot_zsh param, remove glob support
This glob support brings more problem than its is worth it. For
example, there was no support for files with space on its PATH, since
there was a eval call to make glob support works. And the idea of zit is
to be explicit anyway, so removing any magic is better.