travis: Drop ZSH versions 5.0.8 and 5.1.1, add newer versions
ZSH versions in Ubuntu:
- Ubuntu 14.04 uses ZSH 5.0.2
- Ubuntu 16.04 uses ZSH 5.1.1
- Ubuntu 18.04 uses ZSH 5.4.2

The only real problem of dropping ZSH 5.1.1 is because this is a ZSH
version still in use by old Ubuntu LTS (Ubuntu 14.04 can be considered
an ancient version nowadays).

This shouldn't be a hard drop though, it is simply motived because some
tests are broken in CI in those older versions. I can accept patches
fixing this if they do not complicate the logic too much.

Also, add newer versions of ZSH (5.6.2 and 5.7.1). They use the newer
.tar.xz compression instead of .tar.gz, so we need a hack to make it
work in both cases.