zit: Add ZIT_DISABLE_UPGRADE in place of upgrade flag
The old upgrade flag was strange: one extra non-descriptive hidden
parameter included in some functions (zit-in and zit-il) that could be
used to disable upgrade in specific modules. The more zit way to do it
is using a environment variable.

So let's do it: introduce a new ZIT_DISABLE_UPGRADE variable that when
passed disables upgrade for a module. It can also be used to disable
upgrades globally, that may be interesting if the user wants to only
upgrade some specific modules (by unsetting ZIT_DISABLE_UPGRADE).