zit: Rename zit-update to zit-upgrade
This rename is mostly to maintain consistency with the rest of the
code. There is `ZIT_MODULES_UPGRADE` and `ZIT_DISABLE_UPGRADE`, however
the command itself was `zit-update`, making things confusing.

`zit-update` is aliased to `zit-upgrade` for compatibility purposes.
It is a soft-deprecated command. "Soft" because I don't think that I
will ever remove this alias, since the difference between `upgrade` and
`update` doesn't make much sense in this context.